ci.nix: make isBuildable work when `p.meta.license` is a list
Currently when `p.meta.license` is a list, `ci.nix` falls back to its
default value of `true` (it's not an attrset, so it can't have a
`.free` attribute). If that list contains an unfree license, the
actions can end up trying to build the package, and getting an error
because `allowUnfree` isn't set.